fix(db): Better organization of the initial query functions
